# docs/dev/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: goadmin Web API
  version: v1
  description: |
    Контракт API для внешнего React-интерфейса.
    UI разворачивается отдельно от goadmin.
    Legacy режим `X-Subject-ID` допускается только для обратной совместимости и должен быть отключен в production.
servers:
  - url: http://127.0.0.1:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
    ErrorResponse:
      type: object
      required: [request_id, error_code, message]
      properties:
        request_id:
          type: string
        error_code:
          type: string
        message:
          type: string
    ExecuteRequest:
      type: object
      required: [module, command]
      additionalProperties: false
      properties:
        module:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9_.-]+$"
        command:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9_.-]+$"
        args:
          type: array
          maxItems: 16
          items:
            type: string
            maxLength: 256
    ExecuteSuccess:
      type: object
      required: [request_id, status]
      properties:
        request_id:
          type: string
        status:
          type: string
        data:
          type: object
          additionalProperties: true
        error_code:
          type: string
    MetricLatestResponse:
      type: object
      required: [request_id, module, ts, payload]
      properties:
        request_id:
          type: string
        module:
          type: string
        ts:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    AuditQueryResponse:
      type: object
      required: [request_id, items]
      properties:
        request_id:
          type: string
        items:
          type: array
          items:
            type: object
            required: [subject, action, source, status, request_id, ts]
            properties:
              subject:
                type: string
              action:
                type: string
              source:
                type: string
              status:
                type: string
              request_id:
                type: string
              payload:
                type: object
                additionalProperties: true
              ts:
                type: string
                format: date-time
    MeResponse:
      type: object
      required: [request_id, subject, roles, auth_method]
      properties:
        request_id:
          type: string
        subject:
          type: string
        roles:
          type: array
          items:
            type: string
        auth_method:
          type: string
    ModulesResponse:
      type: object
      required: [request_id, items]
      properties:
        request_id:
          type: string
        items:
          type: array
          items:
            type: string
paths:
  /v1/health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
  /v1/me:
    get:
      summary: Current web subject profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current subject
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/modules:
    get:
      summary: List available modules
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Modules list
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModulesResponse"
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/commands/execute:
    post:
      summary: Execute module command
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
      responses:
        "200":
          description: Command executed
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteSuccess"
        "400":
          description: Invalid command payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/metrics/latest:
    get:
      summary: Get latest module metric
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: module
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest metric
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricLatestResponse"
        "400":
          description: Bad query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Metric not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/audit:
    get:
      summary: Query audit events
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: subject
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Audit list
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditQueryResponse"
        "400":
          description: Bad query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"